块同步，分叉处理时一次请求的块数量动态变化，适应网络情况

+++ src/core/groupchain.go
@@ -166,7 +166,7 @@ func (chain *GroupChain) Close() {
 	chain.groups.Close()
 }
 
-func (chain *GroupChain) GetGroupsAfterHeight(height uint64, limit int64) ([]*types.Group) {
+func (chain *GroupChain) GetGroupsAfterHeight(height uint64, limit int) ([]*types.Group) {
 	return chain.getGroupsAfterHeight(height, limit)
 }
